商品详细页商品介绍接口 商品销量 商品库存取消缓存

+++ RearEnd/SmallProfitMall/src/main/java/cn/itcast/service/impl/OrderServiceImpl.java
@@ -464,7 +464,7 @@ public class OrderServiceImpl implements OrderService {
             productId[i] = shoppingCarts.get(i).getProductId();
             productQuantity[i] = shoppingCarts.get(i).getProductQuantity();
         }
-        List<Integer> productContents = productDetailsDao.findProductInventory(productId);
+        List<Double> productContents = productDetailsDao.findProductInventory(productId);
         Boolean sign = false;
 
         for (int i = 0; i <productContents.size() ; i++) {
